+++ typegraph/core/src/runtimes/prisma/mod.rs
@@ -3,6 +3,7 @@
 
 pub mod context;
 pub mod errors;
+mod id_fields;
 pub mod migration;
 mod model;
 pub mod relationship;
